feat(grpc): Add new blocking and async gRPC clients
This commit rewrites the gRPC client used to connect to the emulator.
The previous implementation was a single class and has been replaced
with a more robust, testable, and flexible architecture.

The new implementation provides two distinct client types built via
`EmulatorGrpcClientBuilder`:

* `BlockingEmulatorGrpcClient`: A simple, synchronous client for
  straightforward scripting or tasks.

* `CallbackEmulatorGrpcClient`: An advanced, asynchronous client
  designed for long-lived applications (like UIs). It manages a
  background thread to provide:
  - Non-blocking `connectAsync()`
  - Automatic liveness monitoring to detect server disconnects.
  - A `connectionStateChanges` event source for reactive state updates.

This wil be used by the ui to track the state of the remote endpoint and
to establish connections without blocking the ui.

Bug: 331192780
Change-Id: I5295e889c846c15112bc18fb2561309de49aeb92