Prefer IPv4 over IPv6 in our adb server
Our adbd (the host side) advertisement
contains only the port part and adbd does
not look for IPv6. If we don't open IPv4
port then adbd refuses to work with us.

It happens that goldfish::async::resolveAddress
returns the IPv6 address first and
LibuvServer::bindAndListen was happy with the
first one.

Bug: 439851730
Change-Id: Ibe00b0e4ef9ea0f6a65b578c0a22a759a6049ceb
Signed-off-by: Roman Kiryanov <rkir@google.com>