Enable audio
This cl handles the `-noaudio` and `-audio backend`
command line options.

The `-noaudio` option explicitly disables the audio
(both input and output).

The `-audio backend` option provides the backend
to use on the QEMU side. It must be enabled in
our QEMU build.

If none of the above options provided, we will
try to detect (if available) the backend to use
and pass it to QEMU. This is NOT implemented.

To enable audio pass the `-audio` option.

Bug: 448177089
Test: -audio pa
Change-Id: Ieabc75a4a21ee4e57b5bd4f4437349e6df12adc5
Signed-off-by: Roman Kiryanov <rkir@google.com>