Prevent a deadlock in LibuvEventLoopImpl::shutdown
The lambda passed to `doPost` lock the `mActiveTimersMutex`
to access `mActiveTimers`. If it happens that
`LibuvTimer::doCancel` destroys the last strong pointer
to `LibuvTimer` then the `~LibuvTimer` will be caelled
with `mActiveTimersMutex` locked.

Internally `~LibuvTimer` calls `removeActiveTimer`
which also needs to lock `mActiveTimersMutex`.

This change delays `~LibuvTimer` until
`mActiveTimersMutex` is unlocked.

Change-Id: Ie7440682439dbd9c3d99752d11b6d0a68086da35
Signed-off-by: Roman Kiryanov <rkir@google.com>