Remove unnecessary test_suites and tag integration tests as "manual"
We discover and run unit tests using ... so test suits are unnecessary
and just cause confusion about which tests are run where.

The integration tests take longer to run and interfere with unit test
deflaking (--runs_per_test=1000).

We will explicitly run the test_suits containing the integration tests
on the buildbots.

Bug: 449212254
Change-Id: I20510b174c3186fa06933139fd39d5caa98d265e