Pass more AVD properties directly to the plugin for use by drivers
This saves duplicating the reading logic of the avd config.ini a second time
and fixes a bug when the content directory is overridden. This also
stops the plugin from trying to write out the hardware-qemu.ini file a
second time.

Instead, only the hardware-qemu.ini file, which is written by the
launcher, is read by the plugin.

Bug: 432262765
Change-Id: I93465c5b81f56280f5bb0e5a66f10cb3165cd539