Rename the bazel rule for the launcher to launcher
It is conventional that the cc_binary name matches the name of the
source file containing the main() function.

This also has the advantage of allowing the final part of the label to
be omitted when calling bazel e.g.
bazel build //hardware/generic/goldfish/emulator/launcher

Also rename the launcher binary in the release zip to "emulator", this matches
the existing release.

Note that references to bazel build ...:goldfish will be updated once
this CL is submitted. Though most folks are already using the :launch_emulator
convenience target.

Change-Id: Ie513a68db5222e59296a20858552de26d1fc755c