Fix Mac boot tests that have been silently failing
They require network GRPC to work so allow network access on the sandbox

This oddly triggers pyc creation at build time and there's a clash because launch_emulator and the tests are creating the same file. Make the tests depend on the launch_emulator py_binary to fix that and rename to launch_emulator.py to match.

Change-Id: Ibdc3d9bbb47cd0f84db24a67b4d3ea024c86d610