fix: Make crash database available in bazel runs
When running the emulator using bazel, the crash database location
was not explicitly set, causing it to end up in an unknown directory.
This made it impossible to retrieve crashes on subsequent runs.

This change detects if the process is running inside bazel (in both
the launcher and upload tool) and sets the
ANDROID_EMU_CRASH_REPORTING_DATABASE environment variable to
"/tmp/crash-report.db" if it is not already defined.

Bug: 456299732
Change-Id: Ibd941d5bbbdda24fb7c1ce88494b5e29741ced69