feat(launcher): Update default system image to 16k page size
This updates the default system images used by the emulator
launcher and tests to a `sdk_gphone16k` build (14345414).

The Bazel build files for `launch_emulator` and `boot_test` are
updated to point to the new `@android_minigbm16k` dependencies.

Additionally, the `emulator_lib.py` script is modified to scan
for both standard and `16k` page size system images, allowing it
to correctly locate the new image directories.

Change-Id: Id6a707fc93f9c9310f02352d711cbaa8ae17a440