fix(hal): Unregister reset handlers to prevent crash on reboot
The `BootPropertiesDevice` and `GuestStatusDevice` HALs would register a
QEMU reset handler upon creation but never unregister it upon
destruction.

If an emulator reset (e.g., a guest reboot) was triggered after
these devices were destroyed, QEMU would attempt to call the stale
handler, leading to a use-after-free and a crash.

This change introduces a new `EmulatorResetCallbacks` struct in a
common `emulator_reset.h` header. This struct holds pointers for both
registering and unregistering a reset handler.

The HAL devices are updated to use this struct and now correctly call
the `do_unregister` callback in their destructors, ensuring the handler
is removed. `avd-info.cpp` is updated to provide the new (BQL-safe)
unregister function when initializing the HALs.

Note: the GuestStatusDevice is constantly recreated the moment
that guest boot complete, so a static shared ptr is used to hold
onto the old one for resetHandler to work on valid object.

also removed -no-reboot to enable adb reboot
and fixed uptime to be reset correctly

Bug: 456299732
Change-Id: If410536e91b6eadc3111155aa2b413c9188388db