adb: check boot complete before connecting guest adbd
There are quite some flakiness when adbd is started
early before boot complete, and typically adb track-devices
just times out.

This cl delays the process by waiting till boot complete.

Test shows that this indeed improved the adb robustness
in 10000 tests on RBE

This is not the cleanest solution to the nasty adb connection
problem, and hopefully we can find the root cause later
and improve it.

TODO: b/456020509

Change-Id: Ied02cbd0628cd1056c598413239035cd7d05c5a6