Stop calling android custom file handling functions from C code
This code is not currently used and can probably be removed but that is
a decision for another CL.

This technically introduces a regression with unicode handling on
Windows but as it's not used and we don't currently handle correctly in Qemu
on Windows it shouldn't cause a problem.

In future CLs, we will resolve the handling in Qemu by setting the
codepage to UTF-8 in metadata and the same change could be applied to
this if we decide to keep it.

Bug: 450547571

Change-Id: I3c79b2ff33deb7b3a8e25a784e56f9c54c16204f